remove eliminateExtremeValues in gat telemeters values getTelemeterValueWithoutAmbientLight

+++ Inc/application/statistiques/statistiques.h
@@ -20,7 +20,7 @@
 /* Functions definitions */
 
 int cStandardError (int statistical_series[], int statistical_serial_length);
-int cVariance (int statistical_series[], int statistical_serial_length);
+float cVariance (int statistical_series[], int statistical_serial_length);
 int cAverage (int statistical_series[], int statistical_serial_length);
 int cStandardDeviation (int statistical_series[], int statistical_serial_length);
 void eliminateExtremeValues (int *statistical_series, int *statistical_serial_length);
